@@ -20,7 +20,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.18.2 <3.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -70,6 +70,20 @@ flutter:
     - images/brain.png
     - images/store.png
     - images/home.png
+    - images/tim.png
+    - images/heart.png
+    - images/timused.png
+    - images/5050.png
+    - images/5050used.png
+    - images/call.png
+    - images/callused.png
+    - images/hoikhangia.png
+    - images/hoikhangiaused.png
+    - images/background.png
+    - images/background_home.png
+    - images/reset.png
+    - images/doicauhoi.png
+    - images/phonecall.png
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
 
